Rename `SingleTPUStrategy`
