+++ src/lightning/fabric/strategies/__init__.py
@@ -19,6 +19,7 @@ from lightning.fabric.strategies.parallel import ParallelStrategy  # noqa: F401
 from lightning.fabric.strategies.registry import _call_register_strategies, _StrategyRegistry
 from lightning.fabric.strategies.single_device import SingleDeviceStrategy  # noqa: F401
 from lightning.fabric.strategies.single_tpu import SingleTPUStrategy  # noqa: F401
+from lightning.fabric.strategies.single_xla import SingleDeviceXLAStrategy  # noqa: F401
 from lightning.fabric.strategies.strategy import Strategy  # noqa: F401
 from lightning.fabric.strategies.xla import XLAStrategy  # noqa: F401
