+++ src/lightning/fabric/strategies/single_tpu.py
@@ -11,54 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Dict, Optional
+from typing import Dict
 
-import torch
-
-from lightning.fabric.accelerators import Accelerator
-from lightning.fabric.accelerators.tpu import _XLA_AVAILABLE
-from lightning.fabric.plugins.io.checkpoint_io import CheckpointIO
-from lightning.fabric.plugins.io.xla import XLACheckpointIO
-from lightning.fabric.plugins.precision import Precision
-from lightning.fabric.strategies.single_device import SingleDeviceStrategy
-from lightning.fabric.utilities.types import _DEVICE
+from lightning.fabric.strategies.single_xla import SingleDeviceXLAStrategy
 
 
-class SingleTPUStrategy(SingleDeviceStrategy):
-    """Strategy for training on a single TPU device."""
+class SingleTPUStrategy(SingleDeviceXLAStrategy):
+    """Legacy class.
 
-    def __init__(
-        self,
-        device: _DEVICE,
-        accelerator: Optional[Accelerator] = None,
-        checkpoint_io: Optional[CheckpointIO] = None,
-        precision: Optional[Precision] = None,
-    ):
-        if not _XLA_AVAILABLE:
-            raise ModuleNotFoundError(str(_XLA_AVAILABLE))
-        if isinstance(device, torch.device):
-            # unwrap the `torch.device` in favor of `xla_device`
-            device = device.index
+    Use `SingleDeviceXLAStrategy` instead.
+    """
 
-        import torch_xla.core.xla_model as xm
-
-        super().__init__(
-            accelerator=accelerator,
-            device=xm.xla_device(device),
-            checkpoint_io=checkpoint_io,
-            precision=precision,
-        )
-
-    @property
-    def checkpoint_io(self) -> CheckpointIO:
-        if self._checkpoint_io is None:
-            self._checkpoint_io = XLACheckpointIO()
-        return self._checkpoint_io
-
-    @checkpoint_io.setter
-    def checkpoint_io(self, io: Optional[CheckpointIO]) -> None:
-        self._checkpoint_io = io
-
-    @classmethod
     def register_strategies(cls, strategy_registry: Dict) -> None:
-        strategy_registry.register("single_tpu", cls, description=f"{cls.__class__.__name__}")
+        # legacy
+        strategy_registry.register("single_tpu", cls, description="Legacy class. Use `single_xla` instead.")
